Add integration tests around template recompilation
This is a stab at recreating a testing framework around the
supervisor, starting with some tests around restart behavior as
triggered by hook and configuration recompilation.

For each test case, we use the `hab-sup` binary to start a testing
Habitat package (using just the minimal amount of files, not a
full-blown .hart package). Then, we connect to the supervisor with a
Butterfly client to inject configuration changes. Finally, we inspect
the contents of files within the package's service directory to verify
the desired behaviors.

There's no Docker, there's no package creation; it's just the
`hab-sup` binary and a few files in a fixtures directory.

(Incidentally, a few of the integration tests in this commit will
fail; they cover the bug that will be fixed in the following commit.)