Restart services if hook files change
Previously, we would only restart a service if a configuration file
changed after being re-templated with new configuration
values. Hook files are also able to be templated, however, but we
never looked to see if they changed, which would merit a restart of
the service as well.

Now, we track whether hook files change, too.

Fixes #2448

Signed-off-by: Christopher Maier <cmaier@chef.io>