Add application/environment to ServiceSpec generated from a Service
By not carrying this data over when generating a `ServiceSpec`, we
would always internally be tracking services with _no_
application/environment set (because it defaults to `None`). However,
all the spec files on disk did have this information.

This, coupled with the fact that we watch our spec directory with very
coarse event resolution (basically, if anything in the directory
changes, go check *all* specs to figure out what needs to be
restarted, etc.) means that anytime a new service was started or
stopped (anything that caused a change in the spec directory!), we'd
end up restarting services that had been started with
application/environment set. Our internal representation had no such
information, but the on-disk representation _did_ have it; that's a
difference, so we restart!

Signed-off-by: Christopher Maier <cmaier@chef.io>