[install][REFACTOR] Encapsulate fully-qualified archive name logic
This logic of grabbing an archive name from an identifier was in
several places in the code. The error being returned in all cases was
not quite accurate; the identifier could have been perfectly valid,
but in these cases, it must be a fully-qualified identifier to be of
any use. We now return a `FullyQualifiedPackageIdentRequired` error.

Signed-off-by: Christopher Maier <cmaier@chef.io>