[install][REFACTOR] Remove src_path argument
This turns out to be completely unnecessary. It was only `Some` if we
were dealing with a local archive file, but was only needed to satisfy
the `fetch_artifact` function, which is never called when dealing with
a local archive file, because by the time we get to `fetch_artifact`,
the archive is already in the local cache.

Signed-off-by: Christopher Maier <cmaier@chef.io>