[REFACTOR] Don't install when loading a spec in the manager
There's no reason to go installing packages when we load a service
spec in the manager. That package should have already been installed
when the spec was added. If you go manually screwing around with
Habitat's files under `/hab` then you're already causing problems for
yourself (also, we don't necessarily carry this level of paranoia to
every single operation we do).

With the refactored, single code path for installing packages, we've
removed the special-case logic for whether or not to update an
"at-once" service (that should be the responsibility of the updater
processes only). By not installing here, but just retrieving the
installed package, we can eliminate extra unnecessary queries to the
Builder service.

Signed-off-by: Christopher Maier <cmaier@chef.io>