Allow a hart file to be installed directly from the cache
We allow users to install packages from a local `.hart` file. As part
of this process, we copy the file into the artifact cache. However, if
a user were to install a package that was already in the cache, the
previous implementation would result in a zero-byte file in the cache,
which isn't of much use to anybody.

While this is probably not a very common usecase, the behavior is
annoying and unexpected enough to warrant fixing. Thus, this commit.

Signed-off-by: Christopher Maier <cmaier@chef.io>