Remove --ignore-target hidden option for `hab pkg install`
This option was only added to facilitate the creation of Bintray
packages for Linux and Windows. Originally we were using `hab` itself
to unpack a `core/hab` `.hart` file in order to extract the relevant
binaries to create a non-Habitat `hab` package for those
platforms. However, taking this approach meant that we needed to
disable the platform validation logic for this one use case, or else
we would need a Linux machine to make Linux releases and a Windows
machine to make Windows releases.

Instead, we can just use `tar` to extract the relevant files directly
from the `.hart` file, eliminating the need to use `hab` for this, and
thus eliminating the `--ignore-target` hidden option altogether, which
simplifies our installer code.

Signed-off-by: Christopher Maier <cmaier@chef.io>