Extract service directory creation
There's starting to be more logic around the setup of this directory,
and some future changes we have planned may add some more; we're
extracting it to keep things easy to reason about.

* Implementation Note

We're not using the `Pkg` struct's directory members here, but
constructing the directory paths from our lower-level `fs` module
instead, even though they're the same.

This is because the service-owned directories are present in Pkg, but
the supervisor-owned directories are not. If we were to add the
supervisor-owned directories to `Pkg` for consistency, we would then
expose those directories in templated content, which could cause
unexpected behaviors and permission-related issues.

To keep `Pkg` focused on templating (an association that still needs
strengthening in the codebase), we simply take references to pieces of
`Pkg` that we currently care about, and construct _all_ directory
paths in a consistent manner.

Signed-off-by: Christopher Maier <cmaier@chef.io>