Simplify Package struct
Really, what we want to do is decouple the serialization of the data
from the structure that we're serializing, since those structures can
be used elsewhere for other purposes.

To do that, we can simply wrap a new type around the original one, and
attach our own template-specific serialization logic to it. This gets
us our template context serialization logic in one easy-to-adapt
place, while avoidign lots of boilerplate and needless memory
allocations. It also gives us a new type that we can expand as needed
in the future.

Signed-off-by: Christopher Maier <cmaier@chef.io>