Add a HookRunner abstraction to execute hooks asynchronously
Due to the elaborate and non-standard-library-based approach to
spawning hooks in Windows, we cannot use `tokio-process` to model hook
execution.

Instead, we will wrap existing `Hook` structs in an `Arc<Mutex<...>`
and wrap those in turn in a `HookRunner`, which will itself turn into
a future that executes the hook on a separate thread, signaling back
via a oneshot channel when it has finished.

As we enable more hooks to run asynchronously, this may become more
streamlined.

Signed-off-by: Christopher Maier <cmaier@chef.io>