Modify post-stop hook to be thread-safe
We wrap the post-stop hook in an `Arc<Mutex<...>` to make it sharable
across threads so it can be executed in a future.

As we further develop our async hook story, we may find better ways to
refactor our hook infrastructure. This appears to the be mimimal
change to run the post-stop hook in a future, though.

Signed-off-by: Christopher Maier <cmaier@chef.io>