Stop the CtlAcceptor future at shutdown
By stopping the CtlAcceptor future when we shutdown, we can prevent
users from issuing commands that may put the system in an unexpected
state while we are in the process of shutting down the system.

Users that try to interact with a Supervisor that is shutting down
will now receive a message notifying them of this fact.

It should be noted that the CtlGateway is _still_ operational, so
there will actually be something that receives the requests, but it
will have nothing to _send_ those requests to for further processing.

The reason for doing this is that it allows us to spawn any shutdown
futures we need and then allow them all to run to completion. At this
point, the main Tokio runtime will be idle, at which point we can shut
it down. Since the CtlAcceptor future is a never ending stream, this
would not be possible without explicitly terminating it.

The CtlGateway is currently running in a separate Tokio reactor in
another thread, so it does not affect our shutdown-on-idle logic
here. When we can convert the CtlGateway to a pure future to spawn on
our main runtime, however, we will need to institute similar shutdown
logic for it, as well.

Signed-off-by: Christopher Maier <cmaier@chef.io>