Encapsulate service shutdown in a future
Continuing to work up from the process level, we first call the
process supervisor-shutdown future, then wrap additional service-level
cleanup around it.

Specifically, we remove the health check data for the service, and
then run the service's `post-stop` hook, if it exists.

******************

Additionally, we remove the `update_package` method from `Service`, in
favor of a cleaner model. `update_package` modified a `Service` struct
in place, according to what has changed in the new version of the
package. Historically, this has been the source of several subtle bugs
because certain aspects of this change were missed.

Going forward, "updating" a service will be a relatively simple
process of stopping the service and then starting it up again. The
newly-created `Service` struct will be exactly what it needs to be.

Signed-off-by: Christopher Maier <cmaier@chef.io>