Update the Manager to shut down services asynchronously
Shutting down services is now asynchronous. This encompasses three
distinct scenarios:

- User stops a service with `hab svc stop`
- Service shuts down for an update
- Service shuts down because the Supervisor itself is shutting down for good

All future spawning is taken care of within the run loop of the
Manager.

Instead of calling an explicit `shutdown` method, we instead break out
of the main loop with a `ShutdownMode` enum variant. Then, based on
this value, we end up shutting down the Supervisor cleanly. To do
this, we first stop the CtlAcceptor future, then gather up and spawn
futures for shutting down all services. Once all futures have
finished, we shut down.

An interesting wrinkle comes out of the choice to simplify the service
upgrade logic. Namely, we need to know when the future has finished in
order to know when to restart the new version of the service.

To do this, we store the path of the service's spec file in a special
`specs_to_reload` vector in the Manager itself as the last thing the
shutdown (rather, "upgrade") future for a service. Then, each time
through the main Supervisor loop, we simply load up any specs we find
in this vector.

Signed-off-by: Christopher Maier <cmaier@chef.io>